+++ packages/encryption/src/Domain/Split/EncryptionTypeSplit.ts
@@ -1,0 +1,6 @@
+import { DecryptedPayloadInterface, EncryptedPayloadInterface } from '@standardnotes/models'
+
+export interface EncryptionTypeSplit<T = EncryptedPayloadInterface | DecryptedPayloadInterface> {
+  rootKeyEncryption?: T[]
+  itemsKeyEncryption?: T[]
+}