+++ packages/encryption/src/Domain/Types/ItemAuthenticatedData.ts
@@ -1,0 +1,6 @@
+import { ProtocolVersion, Uuid } from '@standardnotes/common'
+
+export type ItemAuthenticatedData = {
+  u: Uuid
+  v: ProtocolVersion
+}