+++ packages/web/src/javascripts/Components/ContentListView/ContentListView.tsx
@@ -17,6 +17,9 @@ import { NotesController } from '@/Controllers/NotesController'
 import { AccountMenuController } from '@/Controllers/AccountMenu/AccountMenuController'
 import { ElementIds } from '@/Constants/ElementIDs'
 import ContentListHeader from './Header/ContentListHeader'
+import ResponsivePaneContent from '../ResponsivePane/ResponsivePaneContent'
+import { AppPaneId } from '../ResponsivePane/AppPaneMetadata'
+import { useResponsiveAppPane } from '../ResponsivePane/ResponsivePaneProvider'
 
 type Props = {
   accountMenuController: AccountMenuController
@@ -41,6 +44,8 @@ const ContentListView: FunctionComponent<Props> = ({
   notesController,
   selectionController,
 }) => {
+  const { toggleAppPane } = useResponsiveAppPane()
+
   const itemsViewPanelRef = useRef<HTMLDivElement>(null)
 
   const {
@@ -52,24 +57,23 @@ const ContentListView: FunctionComponent<Props> = ({
     panelWidth,
     renderedItems,
     searchBarElement,
-    selectNextItem,
-    selectPreviousItem,
   } = itemListController
 
-  const { selectedItems } = selectionController
+  const { selectedItems, selectNextItem, selectPreviousItem } = selectionController
 
   const isFilesSmartView = useMemo(
     () => navigationController.selected?.uuid === SystemViewId.Files,
     [navigationController.selected?.uuid],
   )
 
-  const addNewItem = useCallback(() => {
+  const addNewItem = useCallback(async () => {
     if (isFilesSmartView) {
       void filesController.uploadNewFile()
     } else {
-      void createNewNote()
+      await createNewNote()
+      toggleAppPane(AppPaneId.Editor)
     }
-  }, [filesController, createNewNote, isFilesSmartView])
+  }, [isFilesSmartView, filesController, createNewNote, toggleAppPane])
 
   useEffect(() => {
     /**
@@ -82,7 +86,7 @@ const ContentListView: FunctionComponent<Props> = ({
       modifiers: [KeyboardModifier.Meta, KeyboardModifier.Ctrl],
       onKeyDown: (event) => {
         event.preventDefault()
-        addNewItem()
+        void addNewItem()
       },
     })
 
@@ -168,11 +172,11 @@ const ContentListView: FunctionComponent<Props> = ({
   return (
     <div
       id="items-column"
-      className="sn-component section app-column app-column-second"
+      className="sn-component section app-column app-column-second flex flex-col border-b border-solid border-border"
       aria-label={'Notes & Files'}
       ref={itemsViewPanelRef}
     >
-      <div className="content">
+      <ResponsivePaneContent paneId={AppPaneId.Items}>
         <div id="items-title-bar" className="section-title-bar border-b border-solid border-border">
           <div id="items-title-bar-container">
             <ContentListHeader
@@ -204,7 +208,7 @@ const ContentListView: FunctionComponent<Props> = ({
             selectionController={selectionController}
           />
         ) : null}
-      </div>
+      </ResponsivePaneContent>
       {itemsViewPanelRef.current && (
         <PanelResizer
           collapsable={true}