+++ packages/web/src/javascripts/Services/AlertService.ts
@@ -1,5 +1,5 @@
-import { ButtonType, sanitizeHtmlString } from '@standardnotes/snjs'
-import { AlertService } from '@standardnotes/services'
+import { AlertService, ButtonType } from '@standardnotes/services'
+import { sanitizeHtmlString } from '@standardnotes/utils'
 import { SKAlert } from '@standardnotes/styles'
 
 /** @returns a promise resolving to true if the user confirmed, false if they canceled */